# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: VCT Platform API
  version: "1.0.0"
  description: |
    API cho hệ thống quản lý giải đấu Vovinam (VCT).
    Backend generated từ spec này sẽ tạo Go structs và TS types tự động.
  contact:
    name: VCT Tech Team
servers:
  - url: http://localhost:18080/api/v1
    description: Local development
  - url: https://api.vct.vn/api/v1
    description: Production
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Làm mới access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token mới
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
  /auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Lấy thông tin người dùng hiện tại
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Thông tin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Chưa xác thực hoặc token hết hạn
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
      required: [username, password]
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, btc, referee_manager, referee, delegate]
...